requirements.txt
asgiref==3.4.1
Django>=4
django-debug-toolbar==3.2.2
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
python-magic==0.4.24; platform_system == "Linux"
python-magic-bin==0.4.14; platform_system == "Windows"
python-magic-bin==0.4.14; platform_system == "Darwin"
django-simple-history=="3.0.0"